feat: 日志页支持弹出独立窗口，可设置启动时自动打开
- 日志控制台抽取为共享组件 LogConsole，设置页与独立窗口共用
- 工具栏新增“弹出独立窗口”按钮：新开 960x640 日志窗口，已存在则聚焦
- 新增“启动时自动打开日志窗口”开关（localStorage 持久化）
- 独立窗口打开期间，设置页日志区显示“已弹出”占位提示，关窗自动恢复
- Rust 新增 open_log_window / is_log_window_open 命令与 log-window:state 事件
- capabilities 加入 log 窗口
