+++ src-tauri/src/utils/log_bridge.rs
@@ -30,6 +30,59 @@ pub fn get_log_history() -> Vec<LogEntry> {
     buffer.iter().cloned().collect()
 }
 
+/// 打开独立日志窗口；已存在时聚焦（仅桌面端，移动端为 no-op）。
+#[tauri::command]
+pub async fn open_log_window(app: AppHandle) -> Result<(), String> {
+    #[cfg(desktop)]
+    {
+        use tauri::{Manager, WebviewUrl, WebviewWindowBuilder};
+
+        if let Some(win) = app.get_webview_window("log") {
+            let _ = win.unminimize();
+            win.set_focus().map_err(|e| format!("聚焦日志窗口失败: {}", e))?;
+            return Ok(());
+        }
+
+        let win = WebviewWindowBuilder::new(
+            &app,
+            "log",
+            WebviewUrl::App("index.html?window=log".into()),
+        )
+        .title("日志")
+        .inner_size(960.0, 640.0)
+        .min_inner_size(480.0, 320.0)
+        .build()
+        .map_err(|e| format!("创建日志窗口失败: {}", e))?;
+
+        // 广播窗口状态，供设置页切换“已弹出”占位提示
+        let app_for_event = app.clone();
+        win.on_window_event(move |event| {
+            if let tauri::WindowEvent::Destroyed = event {
+                let _ = app_for_event.emit("log-window:state", false);
+            }
+        });
+        let _ = app.emit("log-window:state", true);
+
+        tracing::info!("[LogWindow] 独立日志窗口已创建");
+    }
+    Ok(())
+}
+
+/// 查询独立日志窗口是否已打开（移动端恒为 false）。
+#[tauri::command]
+pub fn is_log_window_open(app: AppHandle) -> bool {
+    #[cfg(desktop)]
+    {
+        use tauri::Manager;
+        app.get_webview_window("log").is_some()
+    }
+    #[cfg(not(desktop))]
+    {
+        let _ = app;
+        false
+    }
+}
+
 #[derive(Clone, serde::Serialize)]
 pub struct LogEntry {
     pub timestamp: String,
