+++ src/router/index.ts
@@ -8,6 +8,7 @@ const ComapionMode = () => import('../components/views/CompanionMode.vue')
 const MainMenu = () => import('../components/views/MainMenu.vue')
 const PetMode = () => import('../components/views/PetMode.vue')
 const Second = () => import('../components/views/Second.vue')
+const LogWindow = () => import('../components/views/LogWindow.vue')
 
 // 1. 定义路由表
 const routes = [
@@ -35,6 +36,11 @@ const routes = [
     path: '/second',
     name: 'Second',
     component: Second,
+  },
+  {
+    path: '/log-window',
+    name: 'LogWindow',
+    component: LogWindow,
   },
 ]
